Continuous Release (#64)
* Change title of README

* Update README.md (#54)

* Test for Android enabled. Run with: $gradlew createDebugCoverageReport (#55)

* VAULTIONIZER-229: Rework deleteFile function. (#56)

* Add deleteFile() method to DAO.

* Add deleteFile method.

* Add tag for worker.

* Change signature of deleteFile().

* Rework deleteFile() function and add refactor usage of ReferenceFileSyncWorker.

* Remove comment.

* Replace getter with public property.

* Add additional test for FileCache.

* Add develop build badge.

* VAULTIONIZER-204: Implement file status overview (#57)

* Add button to indicate file progress.

* Add simple layouts for file status and items.

* Add simple fragment.

* Add simple ViewModel.

* Add **exquisite** extension function.

* Add adapter to display name, step and worker status.

* Upgrade project config.

* Add additional tags to identify worker and file.

* Make Work requests unique with a specific tag.

* Add additional constants for unique work.

* Add ViewModel for FileStatusFragment.

* Add handlers for LiveData changes.

* Change order of statements.

* Add strings for file status button.

* Integrate new Fragment into Navigation graph and rename some labels.

* Style changes to status UI.

* Prune already successful work.

* Inflate correct layout for file status item.

* Integrate new ViewModel with file status button.

* WorkInfo is now observed directly in the ViewModel.

* Add additional string for zero tasks.

* Add UI feedback if no file is being processed instead of a blank view.

* Fix some issues with Iconics.

* Integrate UI feedback.

* WorkInfo is now private and MutableLiveData has an empty list as initial value.

* Integrate new strings into file status button.

* Fix typo.

* Fix linting issue.

* Fixed bugs when changed from sha256 to bCrypt (#59)

* Test for Android enabled. Run with: $gradlew createDebugCoverageReport

* Implemented BCrypt hashing with Tests

* Implemented BCrypt hashing with Tests

* Implemented BCrypt hashing with Tests

* Changed name of the package containing the data classes

* Fixed bugs when changed from sha256 to bCrypt

* Implemented space management/settings. (#58)

* Started implementing space permission settings.

* Moved logic to viewmodel.

* Implemented QR code and added string resources.

* Fixed bugs and added some requests.

* VAULTIONIZER-242 Refactor caching possibilities in repositories. (#62)

* Rename ManagedResult to Resource.

* Add loading state to Resource.

* Add heavily modified NetworkBoundResource.

* Remove login data.

* Make functions suspending.

* Integrate NetworkBoundResource into ReferenceFileRepository for testing.

* Replace flows with NetworkBoundResource.

* dispatchError() now defaults to Resource.Error().

* Remove unused file.

* Integrate NetworkBoundResource into AuthRepository.

* Add debug message for exceptions.

* Integrate NetworkBoundResource into MiscRepository.

* Integrate NetworkBoundResource into FileRepository.

* Adjust test to new constructor.

* Remove method and reduce complexity.

* Introduce type for UnknownError.

* Integrate NetworkBoundResource into SpaceRepository.

* Add timestamp for last successful fetch to constructor.

* Integrate NetworkBoundResource into syncReferenceFile() function.

* Add mapping methods to decrease code duplication.

* Add flow extension to collect first successful resource.

* Adjust workers to new interfaces.

* Remove try catch block.

* Remove NetworkBoundResource because these functions are not used in the ViewModel.

* Remove NetworkBoundResource from announceUpload()

* Flow is now longer needed here.

* Add centralized class to store events for Authentication.

* Adjust AuthEvent.

* Cleanup AuthViewModel to new Resource.Loading() state.

* Reorganize strings.

* Adjust UI to new ViewModel LiveData.

* Trigger SonarCloud

* Reduce "cognitive complexity".

* VAULTIONIZER-225 Integrate encryption into upload functions. (#63)

* Implemented encryption iin upload worker. Encryptor throws exception when key does not exist.

* Implemented encryption iin upload worker. Encryptor throws exception when key does not exist.

* Add method to VNSpace to query current key status.

* Add Dialog for request a key.

* Add fileAlertType as optional property to display dialog.

* Emit cryptographical error if key is missing.

* Add strings for key generation dialog.

* Add simple handler for cryptographical errors.

* Adjust to rebased changes.

* Remove unnecessary attributes and init spaces on ViewModel creation.

* Add hint if a space is missing a cryptographical key.

* Fix duplicate key creating bug.

* Remove unused dialog.

* Set encrypted bytes in request.


* Implemented settings UI. (#61)

* Implemented settings UI.

* Implemented some settings stuff.

* Implemented further settings logic.

* Update LoginFragment.kt

Removed test logic


* VAULTIONIZER-205: Implement file viewer. (#66)

* Add util method to convert boolean to visibility.

* Cleanup.

* Minor Fragment and ViewModel cleanups.

* Add in-memory cache to store decryption results.

* Add equals() and hashCode() methods to VNFile.

* Add simple file decryption worker.

* Fix wrong function call.

* Add method to read file from internal folders.

* Add decryption worker.

* Add simple Content Provider.

* Add decryption of files (DEBUG).

* Remove unused data source class and package.

* Add simple text viewer fragment.

* Integrate Text viewer fragment into FileFragment.

* Make text viewer scrollable.

* Add new states to VNFile.

* File status is now observed in MainActivity.

* State of a file is now adjusted according to the state of the background worker.

* Add additional tags to update status.

* Fix wrong status of progress indicator.

* Implement prototype of image viewer and adjust text viewer to new navigation arguments.

* Add method to delete key by string tag.

* Add .gradle/ to gitignore.

* Remove unused method.

* Implemented QR code reader and CRC32. (#65)

* Implemented QR code reader and CRC32.

* Worked on show auth key, input validation for join space and export and copy functionality for auth keys.

* Changed class instantiation to object use.

* Update strings.xml

Escaped special characters in string resources

* Adjusted to backend refactoring. (#67)

* Adjusted to backend refactoring.

* Replace json auth body with HTTP-Header.

* Fix Reference file download.

* Fix UI bug.

* Remove unused attributes.


* Fix wrong caching strategy for root folder. (#68)

+++ app/src/main/java/com/vaultionizer/vaultapp/repository/impl/FileRepositoryImpl.kt
@@ -12,21 +12,18 @@ import com.vaultionizer.vaultapp.data.model.domain.VNFile
 import com.vaultionizer.vaultapp.data.model.domain.VNSpace
 import com.vaultionizer.vaultapp.data.model.rest.refFile.NetworkElement
 import com.vaultionizer.vaultapp.data.model.rest.refFile.NetworkFolder
-import com.vaultionizer.vaultapp.data.model.rest.refFile.NetworkReferenceFile
 import com.vaultionizer.vaultapp.data.model.rest.request.UploadFileRequest
 import com.vaultionizer.vaultapp.data.model.rest.result.ApiResult
-import com.vaultionizer.vaultapp.data.model.rest.result.ManagedResult
+import com.vaultionizer.vaultapp.data.model.rest.result.Resource
 import com.vaultionizer.vaultapp.repository.FileRepository
 import com.vaultionizer.vaultapp.repository.ReferenceFileRepository
 import com.vaultionizer.vaultapp.repository.SpaceRepository
 import com.vaultionizer.vaultapp.repository.SyncRequestRepository
 import com.vaultionizer.vaultapp.service.FileService
 import com.vaultionizer.vaultapp.util.Constants
+import com.vaultionizer.vaultapp.util.extension.collectSuccess
 import com.vaultionizer.vaultapp.util.getFileName
-import com.vaultionizer.vaultapp.worker.DataEncryptionWorker
-import com.vaultionizer.vaultapp.worker.FileDownloadWorker
-import com.vaultionizer.vaultapp.worker.FileUploadWorker
-import com.vaultionizer.vaultapp.worker.ReferenceFileSyncWorker
+import com.vaultionizer.vaultapp.worker.*
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.*
@@ -59,22 +56,31 @@ class FileRepositoryImpl @Inject constructor(
      */
     private val minimumIdCache = mutableMapOf<Long, Long>()
 
-    override suspend fun getFileTree(space: VNSpace): Flow<ManagedResult<VNFile>> {
+    override suspend fun getFileTree(space: VNSpace): Flow<Resource<VNFile>> {
         return flow {
-            val cache = fileCaches[space.id] ?: FileCache(FileCache.IdCachingStrategy.LOCAL_ID)
-            fileCaches[space.id] = cache
-            cache.getRootFile()?.let {
-                emit(ManagedResult.Success(it))
+            if (!space.isKeyAvailable) {
+                emit(Resource.CryptographicalError)
                 return@flow
             }
 
-            val referenceFile = referenceFileRepository.downloadReferenceFile(space)
-            referenceFile.collect {
+            val cache = fileCaches[space.id] ?: FileCache(FileCache.IdCachingStrategy.LOCAL_ID)
+            fileCaches[space.id] = cache
+            cache.rootFile?.let {
+                emit(Resource.Success(it))
+                return@flow
+            }
+
+            referenceFileRepository.downloadReferenceFile(space).collect {
+                if (it is Resource.Loading) {
+                    return@collect
+                }
+
                 when (it) {
-                    is ManagedResult.Success -> {
+                    is Resource.Success -> {
+                        val referenceFile = it.data
                         minimumIdCache[space.id] = -1
                         val affectedIds = mutableSetOf<Long>()
-                        persistNetworkTree(it.data.elements, space, -1, affectedIds)
+                        persistNetworkTree(referenceFile.elements, space, -1, affectedIds)
 
                         localFileDao.deleteAllFilesOfSpaceExceptWithIds(affectedIds, space.id)
                         val localFiles =
@@ -86,15 +92,21 @@ class FileRepositoryImpl @Inject constructor(
                         //                      1) Query sync requests
                         //                      2) Add new VNFile to parent folder
                         cache.addFile(root)
-                        emit(ManagedResult.Success(root))
+                        emit(Resource.Success(root))
+
+                        fileCaches[space.id] = cache
+                        return@collect
                     }
                     else -> {
-                        emit(ManagedResult.Error((it as ManagedResult.Error).statusCode))
+                        @Suppress("UNCHECKED_CAST")
+                        val convertedError = it as? Resource<VNFile>
+
+                        if (convertedError == null) {
+                            emit(Resource.UnknownError)
+                        } else emit(convertedError)
                     }
                 }
             }
-
-            fileCaches.put(space.id, cache)
         }.flowOn(Dispatchers.IO)
     }
 
@@ -104,7 +116,6 @@ class FileRepositoryImpl @Inject constructor(
         parent: VNFile,
     ) {
         withContext(Dispatchers.IO) {
-            val workManager = WorkManager.getInstance(applicationContext)
             val name = resolveFileNameConflicts(parent, name)
 
             // Create file in DB
@@ -141,29 +152,26 @@ class FileRepositoryImpl @Inject constructor(
             val uploadWorkData = workDataOf(
                 Constants.WORKER_SYNC_REQUEST_ID to uploadRequest.requestId,
             )
-            val refWorkData = workDataOf(
-                Constants.WORKER_SPACE_ID to parent.space.id
-            )
 
             val encryptionWorker =
-                OneTimeWorkRequestBuilder<DataEncryptionWorker>().setInputData(encryptionWorkData)
+                prepareFileWorkerBuilder<DataEncryptionWorker>(vnFile, encryptionWorkData)
                     .addTag(Constants.WORKER_TAG_FILE)
+                    .addTag(Constants.WORKER_TAG_ENCRYPTION)
                     .build()
             val uploadWorker =
-                OneTimeWorkRequestBuilder<FileUploadWorker>().setInputData(uploadWorkData)
-                    .addTag(Constants.WORKER_TAG_FILE).build()
-            val referenceFileSyncWorker =
-                OneTimeWorkRequestBuilder<ReferenceFileSyncWorker>().setInputData(refWorkData)
+                prepareFileWorkerBuilder<FileUploadWorker>(vnFile, uploadWorkData)
+                    .addTag(Constants.WORKER_TAG_FILE)
                     .build()
 
             fileCaches[parent.space.id]?.addFile(vnFile)
-            parent.content?.add(vnFile)
+            parent.content!!.add(vnFile)
 
-            workManager
-                .beginWith(encryptionWorker)
-                .then(uploadWorker)
-                .then(referenceFileSyncWorker)
-                .enqueue()
+            enqueueUniqueFileWork(
+                vnFile,
+                encryptionWorker,
+                uploadWorker,
+                buildReferenceFileWorker(vnFile)
+            )
         }
     }
 
@@ -216,41 +224,40 @@ class FileRepositoryImpl @Inject constructor(
             fileCaches[space.id]?.addFile(folder)
             parent.content?.add(folder)
 
-            val refWorkData = workDataOf(
-                Constants.WORKER_SPACE_ID to space.id
-            )
-            val refWorker =
-                OneTimeWorkRequestBuilder<ReferenceFileSyncWorker>().setInputData(refWorkData)
-                    .build()
-
-            WorkManager.getInstance(applicationContext).enqueue(refWorker)
+            enqueueUniqueFileWork(folder, buildReferenceFileWorker(folder))
         }
     }
 
     override suspend fun downloadFile(file: VNFile) {
         withContext(Dispatchers.IO) {
-            val workManager = WorkManager.getInstance(applicationContext)
             val request = syncRequestService.createDownloadRequest(file)
 
             val downloadWorkData = workDataOf(
                 Constants.WORKER_SYNC_REQUEST_ID to request.requestId
             )
-            val refWorkData = workDataOf(
-                Constants.WORKER_SPACE_ID to file.space.id
-            )
 
             file.state = VNFile.State.DOWNLOADING
 
             val downloadWorker =
-                OneTimeWorkRequestBuilder<FileDownloadWorker>().setInputData(downloadWorkData)
+                prepareFileWorkerBuilder<FileDownloadWorker>(file, downloadWorkData)
                     .addTag(Constants.WORKER_TAG_FILE)
                     .build()
-            val referenceFileSyncWorker =
-                OneTimeWorkRequestBuilder<ReferenceFileSyncWorker>().setInputData(refWorkData)
-                    .build()
 
-            workManager.beginWith(downloadWorker).then(referenceFileSyncWorker).enqueue()
+            enqueueUniqueFileWork(file, downloadWorker)
         }
+    }
+
+    override suspend fun decryptFile(file: VNFile) {
+        val decryptionWorkData = workDataOf(
+            Constants.WORKER_FILE_ID to file.localId
+        )
+        val decryptionWorker =
+            prepareFileWorkerBuilder<DataDecryptionWorker>(file, decryptionWorkData)
+                .addTag(Constants.WORKER_TAG_FILE)
+                .addTag(Constants.WORKER_TAG_DECRYPTION)
+                .build()
+
+        enqueueUniqueFileWork(file, decryptionWorker)
     }
 
     override suspend fun getFile(fileId: Long): VNFile? {
@@ -271,52 +278,36 @@ class FileRepositoryImpl @Inject constructor(
     override fun getFileByRemote(spaceId: Long, fileRemoteId: Long): VNFile? =
         fileCaches[spaceId]?.getFile(fileRemoteId)
 
-    override suspend fun announceUpload(spaceId: Long): Flow<ManagedResult<Long>> {
-        return flow {
-            when (val response = fileService.uploadFile(
-                UploadFileRequest(
-                    1,
-                    (spaceRepository.getSpace(spaceId)
-                        .first() as ManagedResult.Success<VNSpace>).data.remoteId
-                )
-            )) {
-                is ApiResult.Success -> {
-                    emit(ManagedResult.Success(response.data))
-                }
-                is ApiResult.NetworkError -> {
-                    emit(ManagedResult.NetworkError(response.exception))
-                }
-                is ApiResult.Error -> {
-                    emit(ManagedResult.Error(response.statusCode))
-                }
-            }
-        }.flowOn(Dispatchers.IO)
+    override suspend fun announceUpload(spaceId: Long): Long? {
+        val space = spaceRepository.getSpace(spaceId).collectSuccess() ?: return null
+
+        val saveIndex = fileService.uploadFile(
+            UploadFileRequest(1),
+            space.remoteId
+        )
+
+        if (saveIndex !is ApiResult.Success) {
+            return null
+        }
+
+        return saveIndex.data
     }
 
-    /**
-     * TODO(jatsqi): Create background worker for this.
-     */
-    override suspend fun deleteFile(file: VNFile): Flow<ManagedResult<VNFile>> {
-        return flow {
-            if (file.parent != null) {
-                file.parent.content?.remove(file)
-                resyncRefFile(file.space).collect {
-                    when (it) {
-                        is ManagedResult.Success -> {
-                            emit(ManagedResult.Success(file))
-                        }
-                        else -> {
-                            emit(ManagedResult.Error(400)) // TODO(jatsqi): Error handling
-                        }
-                    }
-                }
-            }
+    override suspend fun deleteFile(file: VNFile) {
+        localFileDao.deleteFile(file.localId)
+        fileCaches[file.space.id]?.deleteFile(file)
+
+        if (file.parent != null) {
+            file.parent.content?.remove(file)
         }
+
+        enqueueUniqueFileWork(file, buildReferenceFileWorker(file))
     }
 
     override suspend fun updateFileRemoteId(fileId: Long, remoteId: Long) {
         withContext(Dispatchers.IO) {
             localFileDao.updateFileRemoteId(fileId, remoteId)
+            getFile(fileId)?.remoteId = remoteId
         }
     }
 
@@ -427,28 +418,6 @@ class FileRepositoryImpl @Inject constructor(
         return files[-1]!!
     }
 
-    /**
-     * TODO(jatsqi): Remove and create background worker for this
-     */
-    private suspend fun resyncRefFile(space: VNSpace): Flow<ManagedResult<NetworkReferenceFile>> {
-        return flow {
-            val cache = fileCaches[space.id]
-            if (cache?.getRootFile() == null) {
-                emit(ManagedResult.ConsistencyError)
-                return@flow
-            }
-
-            val root = cache.getRootFile()!!.mapToNetwork() as NetworkFolder
-
-            val referenceFile = NetworkReferenceFile(
-                1,
-                root.content ?: mutableListOf()
-            )
-
-            emit(referenceFileRepository.uploadReferenceFile(referenceFile, space).first())
-        }.flowOn(Dispatchers.IO)
-    }
-
     private fun resolveFileNameConflicts(parent: VNFile, name: String): String {
         val nameSet = parent.content?.map { it.name }?.toSet() ?: return name
 
@@ -466,4 +435,41 @@ class FileRepositoryImpl @Inject constructor(
     }
 
     private fun buildDuplicateFileName(name: String, index: Int) = "($index) $name"
+
+    private fun buildReferenceFileWorker(file: VNFile) =
+        prepareFileWorkerBuilder<ReferenceFileSyncWorker>(
+            file,
+            workDataOf(
+                Constants.WORKER_SPACE_ID to file.space.id
+            )
+        ).addTag(Constants.WORKER_TAG_REFERENCE_FILE).build()
+
+    private fun buildDefaultNetworkConstraints() =
+        Constraints.Builder().setRequiredNetworkType(NetworkType.CONNECTED).build()
+
+    private fun enqueueUniqueFileWork(file: VNFile, vararg workers: OneTimeWorkRequest) {
+        var chain = WorkManager.getInstance(applicationContext).beginUniqueWork(
+            String.format(Constants.WORKER_FILE_UNIQUE_NAME_TEMPLATE, file.localId),
+            ExistingWorkPolicy.KEEP,
+            workers[0]
+        )
+
+        for (i in 1 until workers.size) {
+            chain = chain.then(workers[i])
+        }
+
+        chain.enqueue()
+    }
+
+    private inline fun <reified W : ListenableWorker> prepareFileWorkerBuilder(
+        file: VNFile,
+        inputData: Data
+    ): OneTimeWorkRequest.Builder = OneTimeWorkRequestBuilder<W>()
+        .setInputData(inputData)
+        .setConstraints(buildDefaultNetworkConstraints())
+        .addTag(String.format(Constants.WORKER_TAG_FILE_ID_TEMPLATE, file.localId))
+
+    override suspend fun clearLocalFiles(userId: Long) {
+        localFileDao.deleteAllFilesOfUser(userId)
+    }
 }
